refact: move logic to trait
